@@ -1,57 +1,59 @@
 {
-  "_from": "chalk@^1.1.3",
-  "_id": "chalk@1.1.3",
+  "_from": "chalk@^2.1.0",
+  "_id": "chalk@2.1.0",
   "_inBundle": false,
-  "_integrity": "sha1-qBFcVeSnAv5NFQq9OHKCKn4J/Jg=",
+  "_integrity": "sha512-LUHGS/dge4ujbXMJrnihYMcL4AoOweGnw9Tp3kQuqy1Kx5c1qKjqvMJZ6nVJPMWJtKCTN72ZogH3oeSO9g9rXQ==",
   "_location": "/eslint/chalk",
-  "_phantomChildren": {},
+  "_phantomChildren": {
+    "color-convert": "1.9.0",
+    "has-flag": "2.0.0"
+  },
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "chalk@^1.1.3",
+    "raw": "chalk@^2.1.0",
     "name": "chalk",
     "escapedName": "chalk",
-    "rawSpec": "^1.1.3",
+    "rawSpec": "^2.1.0",
     "saveSpec": null,
-    "fetchSpec": "^1.1.3"
+    "fetchSpec": "^2.1.0"
   },
   "_requiredBy": [
     "/eslint",
-    "/eslint/babel-code-frame",
+    "/eslint/inquirer",
     "/eslint/table"
   ],
-  "_resolved": "https://registry.npmjs.org/chalk/-/chalk-1.1.3.tgz",
-  "_shasum": "a8115c55e4a702fe4d150abd3872822a7e09fc98",
-  "_spec": "chalk@^1.1.3",
-  "_where": "/Users/trott/io.js/tools/eslint-tmp/node_modules/eslint",
+  "_resolved": "https://registry.npmjs.org/chalk/-/chalk-2.1.0.tgz",
+  "_shasum": "ac5becf14fa21b99c6c92ca7a7d7cfd5b17e743e",
+  "_spec": "chalk@^2.1.0",
+  "_where": "/Users/apapirovski/Web/nodejs/tools/eslint-tmp/node_modules/eslint",
   "bugs": {
     "url": "https://github.com/chalk/chalk/issues"
   },
   "bundleDependencies": false,
   "dependencies": {
-    "ansi-styles": "^2.2.1",
-    "escape-string-regexp": "^1.0.2",
-    "has-ansi": "^2.0.0",
-    "strip-ansi": "^3.0.0",
-    "supports-color": "^2.0.0"
+    "ansi-styles": "^3.1.0",
+    "escape-string-regexp": "^1.0.5",
+    "supports-color": "^4.0.0"
   },
   "deprecated": false,
-  "description": "Terminal string styling done right. Much color.",
+  "description": "Terminal string styling done right",
   "devDependencies": {
+    "ava": "*",
     "coveralls": "^2.11.2",
-    "matcha": "^0.6.0",
-    "mocha": "*",
-    "nyc": "^3.0.0",
-    "require-uncached": "^1.0.2",
-    "resolve-from": "^1.0.0",
-    "semver": "^4.3.3",
+    "execa": "^0.7.0",
+    "import-fresh": "^2.0.0",
+    "matcha": "^0.7.0",
+    "nyc": "^11.0.2",
+    "resolve-from": "^3.0.0",
     "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "templates.js"
   ],
   "homepage": "https://github.com/chalk/chalk#readme",
   "keywords": [
@@ -78,23 +80,6 @@
     "text"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "name": "Sindre Sorhus",
-      "email": "sindresorhus@gmail.com",
-      "url": "sindresorhus.com"
-    },
-    {
-      "name": "Joshua Appelman",
-      "email": "jappelman@xebia.com",
-      "url": "jbnicolai.com"
-    },
-    {
-      "name": "JD Ballard",
-      "email": "i.am.qix@gmail.com",
-      "url": "github.com/qix-"
-    }
-  ],
   "name": "chalk",
   "repository": {
     "type": "git",
@@ -102,11 +87,10 @@
   },
   "scripts": {
     "bench": "matcha benchmark.js",
-    "coverage": "nyc npm test && nyc report",
-    "coveralls": "nyc npm test && nyc report --reporter=text-lcov | coveralls",
-    "test": "xo && mocha"
+    "coveralls": "nyc report --reporter=text-lcov | coveralls",
+    "test": "xo && nyc ava"
   },
-  "version": "1.1.3",
+  "version": "2.1.0",
   "xo": {
     "envs": [
       "node",
