@@ -1,8 +1,8 @@
 {
   "_from": "function-bind@^1.0.2",
-  "_id": "function-bind@1.1.0",
+  "_id": "function-bind@1.1.1",
   "_inBundle": false,
-  "_integrity": "sha1-FhdnFMgBeY5Ojyz391KUZ7tKV3E=",
+  "_integrity": "sha512-yIovAzMX49sF8Yl58fSCWJ5svSLuaibPxXQJFLmBObTuCr0Mf1KiPopGM9NiFjiYBCbfaa2Fh6breQ6ANVTI0A==",
   "_location": "/function-bind",
   "_phantomChildren": {},
   "_requested": {
@@ -18,10 +18,10 @@
   "_requiredBy": [
     "/has"
   ],
-  "_resolved": "https://registry.npmjs.org/function-bind/-/function-bind-1.1.0.tgz",
-  "_shasum": "16176714c801798e4e8f2cf7f7529467bb4a5771",
+  "_resolved": "https://registry.npmjs.org/function-bind/-/function-bind-1.1.1.tgz",
+  "_shasum": "a56899d3ea3c9bab874bb9773b7c5ede92f4895d",
   "_spec": "function-bind@^1.0.2",
-  "_where": "/Users/trott/io.js/tools/eslint-tmp/node_modules/eslint/node_modules/has",
+  "_where": "/Users/apapirovski/Web/nodejs/tools/eslint-tmp/node_modules/eslint/node_modules/has",
   "author": {
     "name": "Raynos",
     "email": "raynos2@gmail.com"
@@ -44,11 +44,11 @@
   "deprecated": false,
   "description": "Implementation of Function.prototype.bind",
   "devDependencies": {
-    "@ljharb/eslint-config": "^2.1.0",
+    "@ljharb/eslint-config": "^12.2.1",
     "covert": "^1.1.0",
-    "eslint": "^2.0.0",
-    "jscs": "^2.9.0",
-    "tape": "^4.4.0"
+    "eslint": "^4.5.0",
+    "jscs": "^3.0.7",
+    "tape": "^4.8.0"
   },
   "homepage": "https://github.com/Raynos/function-bind",
   "keywords": [
@@ -57,12 +57,7 @@
     "shim",
     "es5"
   ],
-  "licenses": [
-    {
-      "type": "MIT",
-      "url": "http://github.com/Raynos/function-bind/raw/master/LICENSE"
-    }
-  ],
+  "license": "MIT",
   "main": "index",
   "name": "function-bind",
   "repository": {
@@ -71,11 +66,12 @@
   },
   "scripts": {
     "coverage": "covert test/*.js",
-    "coverage-quiet": "covert test/*.js --quiet",
     "eslint": "eslint *.js */*.js",
     "jscs": "jscs *.js */*.js",
     "lint": "npm run jscs && npm run eslint",
-    "test": "npm run lint && npm run tests-only && npm run coverage-quiet",
+    "posttest": "npm run coverage -- --quiet",
+    "pretest": "npm run lint",
+    "test": "npm run tests-only",
     "tests-only": "node test"
   },
   "testling": {
@@ -94,5 +90,5 @@
       "android-browser/4.2..latest"
     ]
   },
-  "version": "1.1.0"
+  "version": "1.1.1"
 }
