@@ -1,27 +1,27 @@
 {
-  "_from": "slice-ansi@0.0.4",
-  "_id": "slice-ansi@0.0.4",
+  "_from": "slice-ansi@1.0.0",
+  "_id": "slice-ansi@1.0.0",
   "_inBundle": false,
-  "_integrity": "sha1-7b+JA/ZvfOL46v1s7tZeJkyDGzU=",
+  "_integrity": "sha512-POqxBK6Lb3q6s047D/XsDVNPnF9Dl8JSaqe9h9lURl0OdNqy/ujDrOiIHtsqXMGbWWTIomRzAMaTyawAU//Reg==",
   "_location": "/eslint/slice-ansi",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "slice-ansi@0.0.4",
+    "raw": "slice-ansi@1.0.0",
     "name": "slice-ansi",
     "escapedName": "slice-ansi",
-    "rawSpec": "0.0.4",
+    "rawSpec": "1.0.0",
     "saveSpec": null,
-    "fetchSpec": "0.0.4"
+    "fetchSpec": "1.0.0"
   },
   "_requiredBy": [
     "/eslint/table"
   ],
-  "_resolved": "https://registry.npmjs.org/slice-ansi/-/slice-ansi-0.0.4.tgz",
-  "_shasum": "edbf8903f66f7ce2f8eafd6ceed65e264c831b35",
-  "_spec": "slice-ansi@0.0.4",
-  "_where": "/Users/trott/io.js/tools/eslint-tmp/node_modules/eslint/node_modules/table",
+  "_resolved": "https://registry.npmjs.org/slice-ansi/-/slice-ansi-1.0.0.tgz",
+  "_shasum": "044f1a49d8842ff307aad6b505ed178bd950134d",
+  "_spec": "slice-ansi@1.0.0",
+  "_where": "/Users/apapirovski/Web/nodejs/tools/eslint-tmp/node_modules/eslint/node_modules/table",
   "author": {
     "name": "David Caccavella",
     "email": "threedeecee@gmail.com"
@@ -30,17 +30,20 @@
     "url": "https://github.com/chalk/slice-ansi/issues"
   },
   "bundleDependencies": false,
-  "dependencies": {},
+  "dependencies": {
+    "is-fullwidth-code-point": "^2.0.0"
+  },
   "deprecated": false,
   "description": "Slice a string with ANSI escape codes",
   "devDependencies": {
-    "ava": "^0.2.0",
-    "chalk": "^1.1.1",
-    "strip-ansi": "^3.0.0",
+    "ava": "*",
+    "chalk": "^2.0.1",
+    "random-item": "^1.0.0",
+    "strip-ansi": "^4.0.0",
     "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
@@ -70,35 +73,13 @@
     "text"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "name": "David Caccavella",
-      "email": "threedeecee@gmail.com",
-      "url": "github.com/dthree"
-    },
-    {
-      "name": "Sindre Sorhus",
-      "email": "sindresorhus@gmail.com",
-      "url": "sindresorhus.com"
-    },
-    {
-      "name": "Joshua Appelman",
-      "email": "jappelman@xebia.com",
-      "url": "jbnicolai.com"
-    },
-    {
-      "name": "JD Ballard",
-      "email": "i.am.qix@gmail.com",
-      "url": "github.com/qix-"
-    }
-  ],
   "name": "slice-ansi",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/chalk/slice-ansi.git"
   },
   "scripts": {
-    "test": "node test.js && xo"
+    "test": "xo && ava"
   },
-  "version": "0.0.4"
+  "version": "1.0.0"
 }
