@@ -1,27 +1,29 @@
 {
-  "_from": "strip-ansi@^3.0.0",
-  "_id": "strip-ansi@3.0.1",
+  "_from": "strip-ansi@^4.0.0",
+  "_id": "strip-ansi@4.0.0",
   "_inBundle": false,
-  "_integrity": "sha1-ajhfuIU9lS1f8F0Oiq+UJ43GPc8=",
+  "_integrity": "sha1-qEeQIusaw2iocTibY1JixQXuNo8=",
   "_location": "/eslint/strip-ansi",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "strip-ansi@^3.0.0",
+    "raw": "strip-ansi@^4.0.0",
     "name": "strip-ansi",
     "escapedName": "strip-ansi",
-    "rawSpec": "^3.0.0",
+    "rawSpec": "^4.0.0",
     "saveSpec": null,
-    "fetchSpec": "^3.0.0"
+    "fetchSpec": "^4.0.0"
   },
   "_requiredBy": [
-    "/eslint/chalk"
+    "/eslint",
+    "/eslint/inquirer",
+    "/eslint/string-width"
   ],
-  "_resolved": "https://registry.npmjs.org/strip-ansi/-/strip-ansi-3.0.1.tgz",
-  "_shasum": "6a385fb8853d952d5ff05d0e8aaf94278dc63dcf",
-  "_spec": "strip-ansi@^3.0.0",
-  "_where": "/Users/trott/io.js/tools/eslint-tmp/node_modules/eslint/node_modules/chalk",
+  "_resolved": "https://registry.npmjs.org/strip-ansi/-/strip-ansi-4.0.0.tgz",
+  "_shasum": "a8479022eb1ac368a871389b635262c505ee368f",
+  "_spec": "strip-ansi@^4.0.0",
+  "_where": "/Users/apapirovski/Web/nodejs/tools/eslint-tmp/node_modules/eslint",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -32,7 +34,7 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "ansi-regex": "^2.0.0"
+    "ansi-regex": "^3.0.0"
   },
   "deprecated": false,
   "description": "Strip ANSI escape codes",
@@ -41,7 +43,7 @@
     "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
@@ -72,23 +74,6 @@
     "text"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "name": "Sindre Sorhus",
-      "email": "sindresorhus@gmail.com",
-      "url": "sindresorhus.com"
-    },
-    {
-      "name": "Joshua Boy Nicolai Appelman",
-      "email": "joshua@jbna.nl",
-      "url": "jbna.nl"
-    },
-    {
-      "name": "JD Ballard",
-      "email": "i.am.qix@gmail.com",
-      "url": "github.com/qix-"
-    }
-  ],
   "name": "strip-ansi",
   "repository": {
     "type": "git",
@@ -97,5 +82,5 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "3.0.1"
+  "version": "4.0.0"
 }
