@@ -1,13 +1,10 @@
 {
   "_from": "table@^4.0.1",
-  "_id": "table@4.0.1",
+  "_id": "table@4.0.2",
   "_inBundle": false,
-  "_integrity": "sha1-qBFsEz+sLGH0pCCrbN9cTWHw5DU=",
+  "_integrity": "sha512-UUkEAPdSGxtRpiV9ozJ5cMTtYiqz7Ni1OGqLXRCynrvzdtR1p+cfOWe2RJLwvUG8hNanaSRjecIqwOjqeatDsA==",
   "_location": "/eslint/table",
-  "_phantomChildren": {
-    "co": "4.6.0",
-    "json-stable-stringify": "1.0.1"
-  },
+  "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
@@ -21,10 +18,10 @@
   "_requiredBy": [
     "/eslint"
   ],
-  "_resolved": "https://registry.npmjs.org/table/-/table-4.0.1.tgz",
-  "_shasum": "a8116c133fac2c61f4a420ab6cdf5c4d61f0e435",
+  "_resolved": "https://registry.npmjs.org/table/-/table-4.0.2.tgz",
+  "_shasum": "a33447375391e766ad34d3486e6e2aedc84d2e36",
   "_spec": "table@^4.0.1",
-  "_where": "/Users/trott/io.js/tools/eslint-tmp/node_modules/eslint",
+  "_where": "/Users/apapirovski/Web/nodejs/tools/eslint-tmp/node_modules/eslint",
   "author": {
     "name": "Gajus Kuizinas",
     "email": "gajus@gajus.com",
@@ -35,31 +32,31 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "ajv": "^4.7.0",
-    "ajv-keywords": "^1.0.0",
-    "chalk": "^1.1.1",
-    "lodash": "^4.0.0",
-    "slice-ansi": "0.0.4",
-    "string-width": "^2.0.0"
+    "ajv": "^5.2.3",
+    "ajv-keywords": "^2.1.0",
+    "chalk": "^2.1.0",
+    "lodash": "^4.17.4",
+    "slice-ansi": "1.0.0",
+    "string-width": "^2.1.1"
   },
   "deprecated": false,
   "description": "Formats data into a string table.",
   "devDependencies": {
-    "ajv-cli": "^1.1.0",
-    "babel": "^6.5.2",
-    "babel-cli": "^6.14.0",
-    "babel-core": "^6.14.0",
-    "babel-plugin-istanbul": "^2.0.3",
-    "babel-preset-es2015-node4": "^2.1.0",
-    "babel-register": "^6.14.0",
-    "chai": "^3.4.1",
-    "eslint": "^3.5.0",
-    "eslint-config-canonical": "^1.8.6",
-    "gitdown": "^2.4.0",
-    "husky": "^0.11.7",
-    "mocha": "^3.0.2",
-    "nyc": "^8.3.1",
-    "sinon": "^1.17.2"
+    "ajv-cli": "^2.1.0",
+    "babel": "^6.23.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-plugin-istanbul": "^4.1.5",
+    "babel-preset-es2015-node4": "^2.1.1",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
+    "eslint": "^4.7.2",
+    "eslint-config-canonical": "^9.3.1",
+    "gitdown": "^2.5.1",
+    "husky": "^0.14.3",
+    "mocha": "^3.5.3",
+    "nyc": "^11.2.1",
+    "sinon": "^4.0.0"
   },
   "homepage": "https://github.com/gajus/table#readme",
   "keywords": [
@@ -96,5 +93,5 @@
     "prepublish": "NODE_ENV=production npm run build",
     "test": "npm run build && nyc --check-coverage mocha"
   },
-  "version": "4.0.1"
+  "version": "4.0.2"
 }
