@@ -43,10 +43,13 @@ process.once("uncaughtException", err => {
 
     if (typeof err.messageTemplate === "string" && err.messageTemplate.length > 0) {
         const template = lodash.template(fs.readFileSync(path.resolve(__dirname, `../messages/${err.messageTemplate}.txt`), "utf-8"));
+        const pkg = require("../package.json");
 
         console.error("\nOops! Something went wrong! :(");
-        console.error(`\n${template(err.messageData || {})}`);
+        console.error(`\nESLint: ${pkg.version}.\n${template(err.messageData || {})}`);
     } else {
+
+        console.error(err.message);
         console.error(err.stack);
     }
 
