@@ -51,7 +51,8 @@ class CodePath {
         Object.defineProperty(
             this,
             "internal",
-            { value: new CodePathState(new IdGenerator(`${id}_`), onLooped) });
+            { value: new CodePathState(new IdGenerator(`${id}_`), onLooped) }
+        );
 
         // Adds this into `childCodePaths` of `upper`.
         if (upper) {
@@ -205,7 +206,7 @@ class CodePath {
 
                 // Call the callback when the first time.
                 if (!skippedSegment) {
-                    callback.call(this, segment, controller); // eslint-disable-line callback-return
+                    callback.call(this, segment, controller);
                     if (segment === lastSegment) {
                         controller.skip();
                     }
