@@ -67,9 +67,13 @@ module.exports = {
          * @private
          */
         function report(node) {
-            context.report({ node, message: "Identifier '{{name}}' is blacklisted.", data: {
-                name: node.name
-            } });
+            context.report({
+                node,
+                message: "Identifier '{{name}}' is blacklisted.",
+                data: {
+                    name: node.name
+                }
+            });
         }
 
         return {
